send: move layout code to layout.go
