+++ ui/page/send/layout.go
@@ -1,0 +1,312 @@
+package send
+
+import (
+	"fmt"
+	"image/color"
+
+	"gioui.org/layout"
+	"gioui.org/widget"
+	"github.com/planetdecred/godcr/ui/decredmaterial"
+	"github.com/planetdecred/godcr/ui/page/components"
+	"github.com/planetdecred/godcr/ui/values"
+)
+
+func (pg *SendPage) initLayoutWidgets() {
+	pg.pageContainer = layout.List{
+		Axis:      layout.Vertical,
+		Alignment: layout.Middle,
+	}
+
+	pg.clearAllBtn = pg.Theme.Button(new(widget.Clickable), "Clear all fields")
+	pg.txFeeCollapsible = pg.Theme.Collapsible()
+
+	pg.nextButton = pg.Theme.Button(new(widget.Clickable), "Next")
+	pg.nextButton.Background = pg.Theme.Color.InactiveGray
+
+	pg.backButton, pg.infoButton = components.SubpageHeaderButtons(pg.Load)
+	pg.backButton.Icon = pg.Icons.ContentClear
+
+	pg.moreOption = pg.Theme.PlainIconButton(new(widget.Clickable), pg.Icons.NavMoreIcon)
+	pg.moreOption.Color = pg.Theme.Color.Gray3
+	pg.moreOption.Inset = layout.UniformInset(values.MarginPadding0)
+
+	pg.sendToButton = pg.Theme.Button(new(widget.Clickable), "Send to account")
+	pg.sendToButton.TextSize = values.TextSize14
+	pg.sendToButton.Background = color.NRGBA{}
+	pg.sendToButton.Color = pg.Theme.Color.Primary
+	pg.sendToButton.Inset = layout.UniformInset(values.MarginPadding0)
+
+	pg.clearAllBtn.Background = pg.Theme.Color.Surface
+	pg.clearAllBtn.Color = pg.Theme.Color.Text
+	pg.clearAllBtn.Inset = layout.UniformInset(values.MarginPadding15)
+}
+
+func (pg *SendPage) topNav(gtx layout.Context) layout.Dimensions {
+	m := values.MarginPadding20
+	return layout.Flex{}.Layout(gtx,
+		layout.Rigid(func(gtx C) D {
+			return layout.Flex{Axis: layout.Horizontal}.Layout(gtx,
+				layout.Rigid(func(gtx C) D {
+					return pg.backButton.Layout(gtx)
+				}),
+				layout.Rigid(func(gtx C) D {
+					return layout.Inset{Left: m}.Layout(gtx, pg.Theme.H6("Send DCR").Layout)
+				}),
+			)
+		}),
+		layout.Flexed(1, func(gtx C) D {
+			return layout.E.Layout(gtx, func(gtx C) D {
+				return layout.Flex{Axis: layout.Horizontal}.Layout(gtx,
+					layout.Rigid(pg.infoButton.Layout),
+					layout.Rigid(func(gtx C) D {
+						return layout.Inset{Left: m}.Layout(gtx, pg.moreOption.Layout)
+					}),
+				)
+			})
+		}),
+	)
+}
+
+func (pg *SendPage) Layout(gtx layout.Context) layout.Dimensions {
+	pageContent := []func(gtx C) D{
+		func(gtx C) D {
+			return pg.pageSections(gtx, "From", false, func(gtx C) D {
+				return pg.sourceAccountSelector.Layout(gtx)
+			})
+		},
+		func(gtx C) D {
+			return pg.toSection(gtx)
+		},
+		func(gtx C) D {
+			return pg.feeSection(gtx)
+		},
+	}
+
+	dims := layout.Stack{Alignment: layout.S}.Layout(gtx,
+		layout.Expanded(func(gtx C) D {
+			return layout.Stack{Alignment: layout.NE}.Layout(gtx,
+				layout.Expanded(func(gtx C) D {
+					return components.UniformPadding(gtx, func(gtx C) D {
+						return layout.Flex{Axis: layout.Vertical}.Layout(gtx,
+							layout.Rigid(func(gtx C) D {
+								return layout.Inset{Bottom: values.MarginPadding16}.Layout(gtx, func(gtx C) D {
+									return pg.topNav(gtx)
+								})
+							}),
+							layout.Rigid(func(gtx C) D {
+								return layout.Inset{Bottom: values.MarginPadding16}.Layout(gtx, func(gtx C) D {
+									return pg.pageContainer.Layout(gtx, len(pageContent), func(gtx C, i int) D {
+										return layout.Inset{Bottom: values.MarginPadding4, Top: values.MarginPadding4}.Layout(gtx, pageContent[i])
+									})
+								})
+							}),
+						)
+					})
+				}),
+				layout.Stacked(func(gtx C) D {
+					if pg.moreOptionIsOpen {
+						inset := layout.Inset{
+							Top:   values.MarginPadding40,
+							Right: values.MarginPadding20,
+						}
+						return inset.Layout(gtx, func(gtx C) D {
+							border := widget.Border{Color: pg.Theme.Color.Background, CornerRadius: values.MarginPadding5, Width: values.MarginPadding1}
+							return border.Layout(gtx, pg.clearAllBtn.Layout)
+						})
+					}
+					return layout.Dimensions{}
+				}),
+			)
+		}),
+		layout.Stacked(func(gtx C) D {
+			gtx.Constraints.Min.Y = gtx.Constraints.Max.Y
+			return layout.S.Layout(gtx, func(gtx C) D {
+				return layout.Inset{Left: values.MarginPadding1}.Layout(gtx, func(gtx C) D {
+					return pg.balanceSection(gtx)
+				})
+			})
+		}),
+	)
+
+	return dims
+}
+
+func (pg *SendPage) pageSections(gtx layout.Context, title string, showAccountSwitch bool, body layout.Widget) layout.Dimensions {
+	return pg.Theme.Card().Layout(gtx, func(gtx C) D {
+		return layout.UniformInset(values.MarginPadding16).Layout(gtx, func(gtx C) D {
+			return layout.Flex{Axis: layout.Vertical}.Layout(gtx,
+				layout.Rigid(func(gtx C) D {
+					return layout.Flex{Axis: layout.Horizontal}.Layout(gtx,
+						layout.Rigid(func(gtx C) D {
+							inset := layout.Inset{
+								Bottom: values.MarginPadding16,
+							}
+							return inset.Layout(gtx, pg.Theme.Body1(title).Layout)
+						}),
+						layout.Flexed(1, func(gtx C) D {
+							if showAccountSwitch {
+								return layout.E.Layout(gtx, func(gtx C) D {
+									inset := layout.Inset{
+										Top: values.MarginPaddingMinus5,
+									}
+									return inset.Layout(gtx, pg.sendDestination.accountSwitch.Layout)
+								})
+							}
+							return layout.Dimensions{}
+						}),
+					)
+				}),
+				layout.Rigid(body),
+			)
+		})
+	})
+}
+
+func (pg *SendPage) toSection(gtx layout.Context) layout.Dimensions {
+
+	return pg.pageSections(gtx, "To", true, func(gtx C) D {
+		return layout.Flex{Axis: layout.Vertical}.Layout(gtx,
+			layout.Rigid(func(gtx C) D {
+				return layout.Inset{
+					Bottom: values.MarginPadding16,
+				}.Layout(gtx, func(gtx C) D {
+					if !pg.sendDestination.sendToAddress {
+						return pg.sendDestination.destinationAccountSelector.Layout(gtx)
+					}
+					return pg.sendDestination.destinationAddressEditor.Layout(gtx)
+				})
+			}),
+			layout.Rigid(func(gtx C) D {
+				if pg.exchangeRate != -1 {
+					return layout.Flex{Axis: layout.Horizontal}.Layout(gtx,
+						layout.Flexed(0.45, func(gtx C) D {
+							return pg.amount.dcrAmountEditor.Layout(gtx)
+						}),
+						layout.Flexed(0.1, func(gtx C) D {
+							return layout.Center.Layout(gtx, func(gtx C) D {
+								return layout.Inset{Top: values.MarginPadding5}.Layout(gtx, func(gtx C) D {
+									icon := pg.Icons.CurrencySwapIcon
+									icon.Scale = 0.45
+									return icon.Layout(gtx)
+								})
+							})
+						}),
+						layout.Flexed(0.45, func(gtx C) D {
+							return pg.amount.usdAmountEditor.Layout(gtx)
+						}),
+					)
+				}
+				return pg.amount.dcrAmountEditor.Layout(gtx)
+			}),
+		)
+	})
+}
+
+func (pg *SendPage) feeSection(gtx layout.Context) layout.Dimensions {
+	collapsibleHeader := func(gtx C) D {
+		feeText := pg.txFee
+		if pg.exchangeRate != -1 {
+			feeText = fmt.Sprintf("%s (%s)", pg.txFee, pg.txFeeUSD)
+		}
+		return pg.Theme.Body1(feeText).Layout(gtx)
+	}
+
+	collapsibleBody := func(gtx C) D {
+		card := pg.Theme.Card()
+		card.Color = pg.Theme.Color.LightGray
+		inset := layout.Inset{
+			Top: values.MarginPadding10,
+		}
+		return inset.Layout(gtx, func(gtx C) D {
+			return card.Layout(gtx, func(gtx C) D {
+				return layout.UniformInset(values.MarginPadding15).Layout(gtx, func(gtx C) D {
+					return layout.Flex{Axis: layout.Vertical}.Layout(gtx,
+						layout.Rigid(func(gtx C) D {
+							//TODO
+							return pg.contentRow(gtx, "Estimated time", "10 minutes (2 blocks)")
+						}),
+						layout.Rigid(func(gtx C) D {
+							inset := layout.Inset{
+								Top:    values.MarginPadding5,
+								Bottom: values.MarginPadding5,
+							}
+							return inset.Layout(gtx, func(gtx C) D {
+								return pg.contentRow(gtx, "Estimated size", pg.estSignedSize)
+							})
+						}),
+						layout.Rigid(func(gtx C) D {
+							return pg.contentRow(gtx, "Fee rate", "10 atoms/Byte")
+						}),
+					)
+				})
+			})
+		})
+	}
+	inset := layout.Inset{
+		Bottom: values.MarginPadding75,
+	}
+	return inset.Layout(gtx, func(gtx C) D {
+		return pg.pageSections(gtx, "Fee", false, func(gtx C) D {
+			return pg.txFeeCollapsible.Layout(gtx, collapsibleHeader, collapsibleBody)
+		})
+	})
+}
+
+func (pg *SendPage) balanceSection(gtx layout.Context) layout.Dimensions {
+	c := pg.Theme.Card()
+	c.Radius = decredmaterial.CornerRadius{NE: 0, NW: 0, SE: 0, SW: 0}
+	return c.Layout(gtx, func(gtx C) D {
+		return components.UniformPadding(gtx, func(gtx C) D {
+			return layout.Flex{Axis: layout.Horizontal}.Layout(gtx,
+				layout.Flexed(0.6, func(gtx C) D {
+					inset := layout.Inset{
+						Right: values.MarginPadding15,
+					}
+					return inset.Layout(gtx, func(gtx C) D {
+						return layout.Flex{Axis: layout.Vertical}.Layout(gtx,
+							layout.Rigid(func(gtx C) D {
+								inset := layout.Inset{
+									Bottom: values.MarginPadding10,
+								}
+								return inset.Layout(gtx, func(gtx C) D {
+									totalCostText := pg.totalCost
+									if pg.exchangeRate != -1 {
+										totalCostText = fmt.Sprintf("%s (%s)", pg.totalCost, pg.totalCostUSD)
+									}
+									return pg.contentRow(gtx, "Total cost", totalCostText)
+								})
+							}),
+							layout.Rigid(func(gtx C) D {
+								return pg.contentRow(gtx, "Balance after send", pg.balanceAfterSend)
+							}),
+						)
+					})
+				}),
+				layout.Flexed(0.3, func(gtx C) D {
+					pg.nextButton.Inset = layout.Inset{Top: values.MarginPadding15, Bottom: values.MarginPadding15}
+					return pg.nextButton.Layout(gtx)
+				}),
+			)
+		})
+	})
+}
+
+func (pg *SendPage) contentRow(gtx layout.Context, leftValue, rightValue string) layout.Dimensions {
+	return layout.Flex{}.Layout(gtx,
+		layout.Rigid(func(gtx C) D {
+			txt := pg.Theme.Body2(leftValue)
+			txt.Color = pg.Theme.Color.Gray
+			return txt.Layout(gtx)
+		}),
+		layout.Flexed(1, func(gtx C) D {
+			return layout.E.Layout(gtx, func(gtx C) D {
+				return layout.Flex{}.Layout(gtx,
+					layout.Rigid(pg.Theme.Body1(rightValue).Layout),
+					layout.Rigid(func(gtx C) D {
+						return layout.Dimensions{}
+					}),
+				)
+			})
+		}),
+	)
+}
